Move yggdrasilctl command line env to separate mod
